Changed notification constant import
