+++ src/__tests__/notificationDb.test.ts
@@ -1,7 +1,7 @@
-import { NOTIFICATION_TYPES } from '../data/constants';
 import { notificationConfigurationFactory, userFactory } from '../db/factories';
 import { NotificationConfigurations, Notifications, Users } from '../db/models';
 
+import { NOTIFICATION_TYPES } from '../db/models/definitions/constants';
 import './setup.ts';
 
 describe('Notification model tests', () => {
