+++ src/data/resolvers/mutations/tasks.ts
@@ -1,8 +1,8 @@
 import { Tasks } from '../../../db/models';
 import { IOrderInput } from '../../../db/models/definitions/boards';
+import { NOTIFICATION_TYPES } from '../../../db/models/definitions/constants';
 import { ITask } from '../../../db/models/definitions/tasks';
 import { IUserDocument } from '../../../db/models/definitions/users';
-import { NOTIFICATION_TYPES } from '../../constants';
 import { checkPermission } from '../../permissions/wrappers';
 import { itemsChange, manageNotifications, sendNotifications } from '../boardUtils';
