+++ src/data/resolvers/mutations/tickets.ts
@@ -1,8 +1,8 @@
 import { Tickets } from '../../../db/models';
 import { IOrderInput } from '../../../db/models/definitions/boards';
+import { NOTIFICATION_TYPES } from '../../../db/models/definitions/constants';
 import { ITicket } from '../../../db/models/definitions/tickets';
 import { IUserDocument } from '../../../db/models/definitions/users';
-import { NOTIFICATION_TYPES } from '../../constants';
 import { checkPermission } from '../../permissions/wrappers';
 import { itemsChange, manageNotifications, sendNotifications } from '../boardUtils';
